chore(banner): rename alert to banner
