--- src/components/alert/CdrAlert.backstop.js
@@ -1,5 +1,0 @@
-module.exports = [{
-  url: 'http://localhost:3000/#/alert',
-  label: 'Alert',
-  responsive: true,
-}];
