--- src/components/alert/styles/CdrAlert.scss
@@ -1,22 +1,0 @@
-@import '../../../css/settings/index.scss';
-@import './CdrAlert.vars.scss';
-
-.cdr-alert {
-  @include cdr-alert-base-mixin;
-}
-
-.cdr-alert--info {
-  @include cdr-alert-info-mixin;
-}
-
-.cdr-alert--warning {
-  @include cdr-alert-warning-mixin;
-}
-
-.cdr-alert--success {
-  @include cdr-alert-success-mixin;
-}
-
-.cdr-alert--error {
-  @include cdr-alert-error-mixin;
-}
