+++ src/components/banner/CdrBanner.backstop.js
@@ -1,0 +1,5 @@
+module.exports = [{
+  url: 'http://localhost:3000/#/banner',
+  label: 'Banner',
+  responsive: true,
+}];
