+++ src/components/banner/__tests__/CdrBanner.spec.js
@@ -1,14 +1,14 @@
 import { mount } from '@vue/test-utils';
-import CdrAlert from 'componentdir/alert/CdrAlert';
+import CdrBanner from 'componentdir/banner/CdrBanner';
 import IconCheckFill from 'componentdir/icon/comps/check-fill';
-describe('CdrAlert', () => {
+describe('CdrBanner', () => {
   it('matches snapshot', () => {
-    const wrapper = mount(CdrAlert, {
+    const wrapper = mount(CdrBanner, {
       propsData: {
         type: 'info',
       },
       slots: {
-       default: [IconCheckFill, 'hey im an alert'],
+       default: [IconCheckFill, 'hey im a banner'],
       }
     });
     expect(wrapper.element).toMatchSnapshot();
